Fix flag for dropping spurious entries during preprocessing
The old argument was specified as a boolean, but an input on the CLI like "True" or "False" would still be interpreted as a string, and thus be truthy.

To fix this, two separate arguments (drop and keep) were created using the store action, setting a global default to true and making the two arguments mutually exclusive.