Update dataset creation and associated documentation
Remove batching and shuffling of tf dataset outside of generator functions and into main scenario scripts. This makes it possible to not shuffle and/or batch the validation dataset.
Make the np.array version repeat the permutation until there is an equal amount of negative and positive examples.
Update documentation