Streamline tf dataset creation and export sequence dataset
Adds an export option to save a csv with positive and negative examples in aa sequence format.
Rearranges the order of conversion between zipped tuples, numpy arrays and streams, in order to make a copy of arrays with aa sequence information of both positive and negative examples available for export, and to perform the sequence-to-image conversion once on this combined set.