Move length filtering step to earlier point in run scenario
Length filtering is now performed immediately after reading in the positive data. Because of this, and due to the fact that the ControlCDR3Source() class also performs length filtering, the size filter steps can be removed from the dataset generation steps. 

This makes it easier to have consistent behaviour between the neg_shuffle_in_cv and negative shuffle on entire dataset variants. Otherwise the latter would have restricted negatives, but not positives.