Change length calculation for shuffle buffer size
The previous method of calculating the size of the tf dataset resulted in an error when run in the hpc environment.
AttributeError: 'Tensor' object has no attribute 'numpy'
To fix this, the size is now derived from the datastream, multiplying it by two if the negatives are generated within the train/test folds.