Move positive labeling to method
Previously, the addition of a column with positive class labels occurred during the initialization of the VdjdbSource object. However, this caused problems when recycling this code for evaluation scripts, because all class labels would get overwritten in case a dataset with positives and negatives would be used as the input.
To fix this, the addition of positive class labels was moved to a method and it was invoked at the required locations.