Avoid getting stuck during negative generation
When the fold or input set of positives on which to generate negatives through shuffling contains only a single epitope, the generation is impossible and can be skipped and a warning is printed.

If the generation takes too long, i.e. in the case where a single majority class outweighs the sum of all other epitopes, it is also impossible to generate enough negatives if the shuffling mechanism tries to use every cdr3 exactly once. If the generation takes too many tries, a warning is printed and a random batch of new cdr3s is sampled and matched to a negative epitope. These cdr3s will appear multiple times in the negatives instead of having a 1-1 pos-neg cdr3 count.