Update datasets and fix downsampling
Downsampling used the same fraction for the TRA and TRB and mixed datasets, despite the epitope distribution being different. The new makefile preprocessing steps generate datasets that downsample the desired epitopes to the size relevant for the specific TRA/B background.

The counts and relevant bash commands were added to the makefile for documentation purposes.

Unused datasets were removed and commented in the makefile.