Fix error in negative data generation via shuffling on entire dataset
After the changes in commit df8b6ed3cae73007167b921948118afe93384cae, an error was introduced in the negative data shuffling functionality of the vdjdb_source class. Seed was introduced as a variable to feed into the sample_pairs function, but seed was not defined because the enumerate call that should generate it was missing.