Fix negative sampler for external validation set
The negative sampler was only using the provided "full_dataset" when checking for false negatives during negative data generation via shuffling. When an external validation set is used, which is not contained inside this provided full dataset (which would be the case when cross-validation is used), accidental false negatives within the validation set can slip through.

The same thing can happen the other way around, when the full dataset and training set differ. 

To fix this, the full dataset was merged with the provided train/validation set, and this combined dataset is used to check for false negatives.