Allow cdr3 sampling for shuffled negative data
Adds the option (via a CLI argument epitope_ratio) to sample CDR3 sequences for every epitope group, instead of sampling an epitope for every CDR3. 

This new option can ensure that the positive/negative ratio is exactly 50:50 for each individual epitope group, at the expense of 1) certain CDR3 sequences appearing multiple times among the negatives, and 2) certain CDR3 sequences only appearing among the positives. This is a consequence of performing a sampling operation once for each epitope group; the operation becomes sampling with replacement, which results in only about 63% of the data being drawn.

When not using the epitope_ratio option, a valid epitope is sampled for each CDR3 sequence (not unique, but just by iterating over all positive sequence pairs), which ensures that every CDR3 sequence occurs once as a positive and once as a negative. However, abundant epitopes will have a lower proportion of negatives than rare epitopes. The reason is that for the former, there are a large number of CDR3 sequences, for which a different epitope will be drawn. Only non-abundant epitopes are allowed to draw the abundant epitope as their negative, but the absolute number will be lower as a result.

This commit also cleans up the code related to negative data generation via shuffling in general: refactors function names, removes duplicated code segments (e.g. in vdjdb source class) and improves docstrings/comments.