Fix distance color bars
Remove color bars for kfold/non-grouped plots, because distances may differ between iterations, making it impossible to assign a single color value to each box in the boxplot.

Fix distance color bars for evaluate box plot and decoy models. Previously a range of e.g. 1:10, 11:20, 21:30 would be used for 10 epitopes split across 3 types. However, with decoy/TRA evaluation plots not every type has the same number of epitopes. Similarly, for decoy plots, the order is completely different. To fix this, iterate through the ordered list of unique epitopes (= x-axis order). There will be a patch for every epitope-type combination, even if a specific type has no value for that epitope. I.e. the length of the patch list equals the number of types times the number of epitopes. To make sure that the correct patch can be retrieved and coloured for the given epitope (times the number of epitopes + current type, as an index, 1-10, 11-20, etc.), the order of the types (=hue) is fixed, and the dataframe is sorted on this when looping through the patch recolouring step. 

This is important, because when sorting the dataframe on [epitope,type], the order of the types is read from this dataframe in order of first appearance, and because not every type has all epitopes, the first epitope might be a decoy one, resulting in an order of [decoy, true], whereas when looping through the epitopes/type combination during patch recolouring, previously the order was determined by an individual sort of both, which would be different (i.e. alphabetical types).

Moves sorting step to start of function, so that all later steps are using the same sorted dataframe. 

Draw multiple colour bar + ticks in case the different types differ in distance range.